pyproject.toml
[project]
name = "vscode-open"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "psutil>=7.1.2",
    "pyobjc>=12.0",
    "pyobjc-framework-scriptingbridge>=12.0",
]
